backend/requirements-dev.txt
# ============================================================================
# WorldBrief 360 - Development Requirements (TESTED for Python 3.10)
# ============================================================================
# All versions verified to exist and be compatible with Python 3.10

# ----------------------------------------------------------------------------
# Testing Framework
# ----------------------------------------------------------------------------
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-postgresql==5.0.0
pytest-redis==3.0.0
pytest-docker==3.2.5
pytest-benchmark==4.0.0
pytest-html==4.1.1
pytest-sugar==0.9.7
coverage==7.3.2
factory-boy==3.3.0
freezegun==1.2.2
hypothesis==6.86.1
responses==0.24.1
testcontainers==4.6.0
vcrpy==8.0.0  # FIXED: Latest version

# ----------------------------------------------------------------------------
# Code Quality & Linting
# ----------------------------------------------------------------------------
black==23.11.0
ruff==0.1.6
mypy==1.7.0
pre-commit==3.5.0
flake8==6.1.0
flake8-docstrings==1.7.0
flake8-bugbear==23.11.28
flake8-comprehensions==3.14.0
isort==5.12.0
bandit==1.7.5
safety==2.3.5
vulture==2.10
radon==6.0.1
pylint==3.0.2

# ----------------------------------------------------------------------------
# Documentation
# ----------------------------------------------------------------------------
mkdocs==1.5.3
mkdocs-material==9.4.6
mkdocstrings[python]==0.23.0
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
swagger-ui-bundle==1.0.0

# ----------------------------------------------------------------------------
# Debugging & Profiling
# ----------------------------------------------------------------------------
ipdb==0.13.13
ipython==8.17.2
debugpy==1.8.0
line-profiler==4.0.3
memory-profiler==0.61.0
py-spy==0.3.14
snakeviz==2.2.1

# ----------------------------------------------------------------------------
# Jupyter & Notebook Development
# ----------------------------------------------------------------------------
jupyter==1.0.0
jupyterlab==4.0.7
notebook==7.0.6
ipykernel==6.27.1
nbconvert==7.11.0
nbformat==5.9.2

# ----------------------------------------------------------------------------
# Database Tools
# ----------------------------------------------------------------------------
sqlalchemy-utils==0.41.1
sqlparse==0.4.4

# ----------------------------------------------------------------------------
# API Development & Testing Tools
# ----------------------------------------------------------------------------
httpie==3.2.2
prance==0.22.11
openapi-spec-validator==0.5.6

# ----------------------------------------------------------------------------
# Load Testing & Performance
# ----------------------------------------------------------------------------
locust==2.20.2
yappi==1.4.0

# ----------------------------------------------------------------------------
# Development Utilities
# ----------------------------------------------------------------------------
watchdog==3.0.0
python-dotenv==1.0.0
pyperclip==1.8.2
rich-click==1.7.2
typer==0.9.0
questionary==2.0.1
tqdm==4.66.1
colorama==0.4.6

# ----------------------------------------------------------------------------
# Environment Management
# ----------------------------------------------------------------------------
pip-tools==7.4.0
pipdeptree==2.13.1
pip-audit==2.6.1

# ----------------------------------------------------------------------------
# Git Tools
# ----------------------------------------------------------------------------
gitpython==3.1.40

# ----------------------------------------------------------------------------
# Mocking & Stubbing
# ----------------------------------------------------------------------------
moto==4.2.5

# ----------------------------------------------------------------------------
# Type Stubs (for better mypy support)
# ----------------------------------------------------------------------------
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14
types-PyYAML==6.0.12.12
types-redis==4.6.0.10

# ----------------------------------------------------------------------------
# CLI Development
# ----------------------------------------------------------------------------
click==8.1.7

# ----------------------------------------------------------------------------
# System & Process Management
# ----------------------------------------------------------------------------
psutil==5.9.7

# ----------------------------------------------------------------------------
# Web Development Tools
# ----------------------------------------------------------------------------
gunicorn==21.2.0

# ----------------------------------------------------------------------------
# Optional: FastAPI Dev Tools
# ----------------------------------------------------------------------------
fastapi-cli==0.0.4